Createduser.js removed
